Add CMake support Move main function to main.cpp Move SBFspot types to Types.h
